fix: fix that make crash delete method when a null image is send to delete the image in data base

+++ app/Helpers/ImagesFunctions.php
@@ -19,10 +19,12 @@ if (!function_exists('deleteLoadedImage')) {
     function deleteLoadedImage($imageLoaded)
     {
 
-        $isThereAnImage = Storage::disk('public')->exists($imageLoaded);
+        if (!is_null($imageLoaded)) {
+            $isThereAnImage = Storage::disk('public')->exists($imageLoaded);
 
-        if ($isThereAnImage) {
-            Storage::disk('public')->delete($imageLoaded);
+            if ($isThereAnImage) {
+                Storage::disk('public')->delete($imageLoaded);
+            }
         }
     }
 }
